Hilt setup working and some main files

+++ client-app/app/src/main/java/com/feup/jtp/client_app/domain/model/Product.kt
@@ -3,5 +3,5 @@ package com.feup.jtp.client_app.domain.model
 data class Product(
     val id: String,
     val name: String,
-    val price: Float
+    val price: Double
 )
